Check for version when resolving dependency notations (#19)
Previously we checked for the JxBrowser version right after the plugin was applied and could see the ISE in the console at this stage. Because of this, we can't configure the plugin extension with IDEA hints, as everything breaks at the stage of applying the plugin.

In this PR, we do JxBrowser version checking while resolving dependency notations, not when applying the plugin.