Support Windows ARM (#54)
In this changeset, we add a dependency to Windows ARM binaries since we
recently
[introduced](https://github.com/TeamDev-IP/JxBrowser/issues/2990) the
support of this architecture. As for now, this will only be supported in
JxBrowser 8.

Also, I added a check for some artifacts supported only in JxBrowser 8.
From now on, if the user decides to use the `jxbrowser-kotlin` artifact
with JxBrowser 7, for example, the build will fail.