Make `version` a lazily evaluated property (#71)
In this changeset, we allow configuring JxBrowser version with a [lazily
evaluated
value](https://docs.gradle.org/current/userguide/properties_providers.html)
by changing `version` field to be a lazy `Property`.

## Problem

The Gradle [documentation
states](https://docs.gradle.org/current/userguide/properties_providers.html):

> When implementing a custom task or plugin, it’s imperative that you
use these lazy properties.

We use regular `String` properties instead. That is OK when a version is
a string literal in `build.gradle.kts`. But it's not OK, when the value
is taken from elsewhere. For example, from the version catalog:

```
jxbrowser {
    // Calling `get()` this way is a bad practice, because the value may not be available yet. 
    version = libs.versions.jxbrowser.asProvider().get()
}
```

## Solution

We change the `version` field from `String` to `Property<String>`. We
don't change other fields because they're less likely to come from the
lazily evaluated sources. And we want to avoid breaking changes. More on
this in the next section.

## Backward compatibility

The plug-in stays backward compatible for Gradle 8.2 and newer. No
changes are required, because of
[`Property::assign`](https://github.com/gradle/gradle/blob/fcef0ec368956438f2c1272b64bad82c996ab236/platforms/core-configuration/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/PropertyExtensions.kt#L41)
extension method.

Users of older Gradle versions, will need to replace `version = "8.2.2"`
with `version.set("8.2.2")`. Otherwise, the build will fail.

## Version bump

Since this change is breaking for some of the users, we bump the version
to `2.0.0`.