Use Java 17 for plugin publishing (#98)
This PR updates the Java version to 17 in the GitHub Actions script used
for publishing the plugin.
Gradle 9 requires this version to work:
https://gradle.org/whats-new/gradle-9/#java-versions-and-toolchains

Also, we no longer need a separate test for verifying the download of
the JxBrowser 8.0.0 jars.
The full list of jars should be tested when updating the JxBrowser
version.